+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## unreleased 1.8.0-RC3
 
 - Portal login with registration in validation
+- Company Certificate
+  - Provide new link from user menu for the company certificate page
+  - UI components created to display filter, sort and company certificates
 - Company Roles
   - Fetch the standard library data from standards.json and implement Table component to display it in the company roles section.
-  - 
 
 ## 1.8.0-RC2